+++ src/main/java/com/seacroak/plushables/block/DjungelskogBlock.java
@@ -1,48 +1,15 @@
 package com.seacroak.plushables.block;
 
-import com.seacroak.plushables.util.VoxelShapeUtils;
-import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
-import net.minecraft.block.*;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemPlacementContext;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.sound.BlockSoundGroup;
-import net.minecraft.state.StateManager;
-import net.minecraft.state.property.Properties;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
-import net.minecraft.world.BlockView;
-import net.minecraft.world.World;
-
-import java.util.Random;
 
 public class DjungelskogBlock extends SimplePlushable {
-    public static Random rand;
+  public DjungelskogBlock() {
+    super();
+  }
 
-    public DjungelskogBlock() {
-        super();
-        rand = new Random();
-    }
-
-    @Override
-    public BlockRenderType getRenderType(BlockState state) {
-        return BlockRenderType.MODEL;
-    }
-
-    @Override
-    public BlockState getPlacementState(ItemPlacementContext context) {
-        return this.getDefaultState().with(Properties.HORIZONTAL_FACING, context.getHorizontalPlayerFacing().getOpposite());
-
-    }
-  static final VoxelShape blockShape = getShape();
-  static final VoxelShape[] blockShapes = { blockShape,
-    VoxelShapeUtils.rotateShape(Direction.NORTH, Direction.EAST, blockShape),
-    VoxelShapeUtils.rotateShape(Direction.NORTH, Direction.SOUTH, blockShape),
-    VoxelShapeUtils.rotateShape(Direction.NORTH, Direction.WEST, blockShape) };
-
-  static public VoxelShape getShape() {
+  @Override
+  public VoxelShape getShape() {
     VoxelShape shape = VoxelShapes.empty();
     shape = VoxelShapes.union(shape, VoxelShapes.cuboid(0.1875, 0, 0.0625, 0.3125, 0.125, 0.25));
     shape = VoxelShapes.union(shape, VoxelShapes.cuboid(0.6875, 0, 0.0625, 0.8125, 0.125, 0.25));
@@ -55,32 +22,4 @@ public class DjungelskogBlock extends SimplePlushable {
     shape = VoxelShapes.union(shape, VoxelShapes.cuboid(0.4375, 0, 0.875, 0.5625, 0.125, 0.9375));
     return shape;
   }
-
-  @Override
-  public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-    Direction direction = (Direction) state.get(FACING);
-
-    switch (direction) {
-      case NORTH: {
-        return blockShapes[0];
-      }
-      case EAST: {
-        return blockShapes[1];
-      }
-      case SOUTH: {
-        return blockShapes[2];
-
-      }
-      case WEST: {
-        return blockShapes[3];
-
-      }
-      default:
-        return blockShape;
-    }
-  }
-    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
-        builder.add(FACING);
-    }
-
 }
