+++ src/main/resources/assets/plushables/blockstates/rattiam_plushable.json
@@ -1,0 +1,20 @@
+{
+  "variants": {
+    "facing=north": {
+      "y": 0,
+      "model": "plushables:block/rattiam_plushable"
+    },
+    "facing=south": {
+      "y": 180,
+      "model": "plushables:block/rattiam_plushable"
+    },
+    "facing=east": {
+      "y": 90,
+      "model": "plushables:block/rattiam_plushable"
+    },
+    "facing=west": {
+      "y": 270,
+      "model": "plushables:block/rattiam_plushable"
+    }
+  }
+}
