+++ src/main/resources/assets/plushables/lang/en_us.json
@@ -8,5 +8,7 @@
   "block.plushables.clucky_plushable": "Plushable §lClucky",
   "block.plushables.pig_plushable": "Plushable §c§lPig",
   "block.plushables.truffles_plushable": "Plushable §c§lTruffles",
-  "block.plushables.djungelskog_plushable": "Plushable §6§lDjungelskog"
+  "block.plushables.djungelskog_plushable": "Plushable §6§lDjungelskog",
+  "block.plushables.rattiam_plushable": "Plushable §7§lRattiam"
+
 }
