+++ src/main/resources/assets/plushables/models/item/rattiam_plushable.json
@@ -1,0 +1,4 @@
+{
+  "parent": "plushables:block/rattiam_plushable",
+  "gui_light": "front"
+}
